+++ src/app/modules/map/components/feature-ais.component.ts
@@ -261,6 +261,7 @@ export class AisVesselsComponent implements OnInit, OnDestroy, OnChanges {
     setTargetStyle(id: any) {
         if(!id) { return }
         let target= this.aisTargets.get(id);
+        if(!target) { return }
         let label= this.formatlabel( target.name || target.callsign || target.mmsi || '');
         let fstyle: any;
         // ** check if stale